requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Template Engine
jinja2==3.1.3

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1

# Recipe Scraping
recipe-scrapers==14.51.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Markdown & File Processing
python-frontmatter==1.1.0
markdown==3.5.2
python-slugify==8.0.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Date & Time
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
ruff==0.1.14

# Logging
python-json-logger==2.0.7
